user models testing working
