+++ tests/server/routes/order.routes.spec.js
@@ -4,7 +4,7 @@ var expect = require('chai').expect;
 var Promise = require('bluebird');
 var order = require('../../../server/app/routes/order.js');
 
-describe('Order Route:', function () {
+xdescribe('Order Route:', function () {
 
 
  /**
